+++ packages/config/src/index.ts
@@ -1,5 +1,0 @@
-export {Runtime} from '@sewing-kit/core';
-export {createPackage} from './package';
-export {createService} from './service';
-export {createWebApp} from './web-app';
-export {createWorkspace} from './workspace';