+++ packages/core/src/core/index.ts
@@ -1,0 +1,34 @@
+export {Runtime, ProjectKind} from './types';
+export {Package, PackageBinary, PackageEntry} from './package';
+export type {
+  PackageBinaryOptions,
+  PackageEntryOptions,
+  PackageOptions,
+} from './package';
+export {WebApp, ServiceWorker} from './web-app';
+export type {WebAppOptions, ServiceWorkerOptions} from './web-app';
+export {Service} from './service';
+export type {ServiceOptions} from './service';
+export {Workspace} from './workspace';
+export type {WorkspaceOptions} from './workspace';
+
+export type Project =
+  | import('./package').Package
+  | import('./web-app').WebApp
+  | import('./service').Service;
+
+export type {
+  FileSystem,
+  Log,
+  LogFormatter,
+  LogOptions,
+  Loggable,
+  Step,
+  StepResources,
+  StepRunner,
+  StepStdio,
+} from './types';
+export {LogLevel} from './types';
+export {DiagnosticError, isDiagnosticError} from './errors';
+export {TargetBuilder, TargetRuntime} from './target';
+export type {Target} from './target';