+++ packages/core/src/plugins/index.ts
@@ -1,0 +1,34 @@
+export type {PluginApi} from './api';
+export {
+  PLUGIN_MARKER,
+  PluginTarget,
+  createComposedProjectPlugin,
+  createComposedWorkspacePlugin,
+  createProjectBuildPlugin,
+  createProjectDevPlugin,
+  createProjectPlugin,
+  createProjectTestPlugin,
+  createWorkspaceBuildPlugin,
+  createWorkspaceDevPlugin,
+  createWorkspaceLintPlugin,
+  createWorkspacePlugin,
+  createWorkspaceTestPlugin,
+  createWorkspaceTypeCheckPlugin,
+} from './plugins';
+export type {
+  AnyPlugin,
+  PluginComposer,
+  ProjectPlugin,
+  ProjectPluginContext,
+  WorkspacePlugin,
+  WorkspacePluginContext,
+} from './plugins';
+export {
+  toArgs,
+  addHooks,
+  projectTypeSwitch,
+  unwrapPossibleGetter,
+  unwrapPossibleArrayGetter,
+} from './utilities';
+export type {ValueOrArray, ValueOrGetter} from './utilities';
+export {MissingPluginError} from './errors';